requirements.txt
langchain[openai]
streamlit
httpx
python-dotenv
streamlit-float